Added auth to customer page, and moved the form handler to another component
