+++ client/src/components/CustomerReviews.js
@@ -1,0 +1,32 @@
+import Auth from '../utils/auth';
+const CustomerReviews = () => {
+     //For the ADD_REVIEW:
+    //const [newReview, setNewReview] = useState();
+    //const [addReview, {error, loading}] = useMutation(ADD_REVIEW);
+    const handleSavedReview = () => {
+        const token = Auth.loggedIn() ? Auth.getToken() : null;    
+        if (!token) {
+          return false;
+        }
+        try{
+
+           // const {data} = await addReview({
+            //    variables: {}
+           // })
+
+        } catch (err){
+            console.error(err);
+        }
+
+    }
+    return (
+        <form>
+        <label>Leave a review:</label>
+        <textarea name='review' rows="4" cols="50"/>
+       </form>
+
+
+      );
+}
+ 
+export default CustomerReviews;
